bigint: Move some parsing from `BoxedLimbs` to `OwnedModulusValue`.
Take a step towards removing all parsing from `BoxedLimbs`.

This has the added benefit of rejecting more invalid modulus values,
notably too-large ones, before allocating memory for them.
