+++ src/arithmetic/bigint/private_exponent.rs
@@ -43,20 +43,27 @@ impl PrivateExponent {
         })
     }
 
+    // Create a `PrivateExponent` with a value that we do not support in
+    // production use, to allow testing with additional test vectors.
     #[cfg(test)]
     pub fn from_be_bytes_for_test_only<M>(
         input: untrusted::Input,
         p: &Modulus<M>,
     ) -> Result<Self, error::Unspecified> {
+        use crate::limb::LIMB_BYTES;
+
         // Do exactly what `from_be_bytes_padded` does for any inputs it accepts.
         if let r @ Ok(_) = Self::from_be_bytes_padded(input, p) {
             return r;
         }
 
-        let dP = BoxedLimbs::<M>::positive_minimal_width_from_be_bytes(input)?;
+        let num_limbs = (input.len() + LIMB_BYTES - 1) / LIMB_BYTES;
+        let mut limbs = BoxedLimbs::<M>::zero(num_limbs);
+        limb::parse_big_endian_and_pad_consttime(input, &mut limbs)
+            .map_err(|error::Unspecified| error::KeyRejected::unexpected_error())?;
 
         Ok(Self {
-            limbs: dP.into_limbs(),
+            limbs: limbs.into_limbs(),
         })
     }
